+++ packages/flutter_meedu/lib/flutter_meedu.dart
@@ -3,6 +3,7 @@ library flutter_meedu;
 export 'meedu.dart';
 export 'page.dart';
 export 'router.dart';
+export 'navigator.dart';
 export 'rx.dart';
 export 'screen_utils.dart';
 export 'state.dart';