+++ packages/flutter_meedu/lib/navigator.dart
@@ -1,5 +1,5 @@
 library navigator;
 
 export 'src/navigation/navigator/navigator.dart';
-export 'src/navigation/navigator/navigator_mixin.dart';
+export 'src/navigation/navigator/navigator_extension.dart';
 export 'src/navigation/transitions/transition.dart';