+++ packages/flutter_meedu/lib/router.dart
@@ -1,5 +1,4 @@
 library router;
 
 export 'src/navigation/navigator_observer.dart';
-export 'src/navigation/router/router.dart';
-export 'src/navigation/router/utils.dart';
+export 'src/navigation/router/contextless_navigator.dart';