+++ packages/flutter_meedu/lib/src/navigation/navigator_observer.dart
@@ -5,11 +5,8 @@ import 'package:flutter/widgets.dart';
 import '../../meedu.dart';
 import 'router/contextless_navigator.dart';
 
-/// the observer to listen the changes in the stack route
-NavigatorObserver get observer => _NavigatorObserver();
-
 /// this class is used to listen the changed in the stack route
-class _NavigatorObserver extends NavigatorObserver {
+class MeeduNavigatorObserver extends NavigatorObserver {
   /// return a name for a route
   String _getRouteName(PageRoute route) {
     return '${route.hashCode}';