+++ packages/flutter_meedu/test/dependency_injection/dependency_injection_test.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:flutter_meedu/router.dart' as router;
-import 'package:meedu/get.dart';
+import 'package:flutter_meedu/flutter_meedu.dart';
 
 void main() {
   testWidgets(