+++ packages/flutter_meedu/test/listeners/provider_listener_state_provider_with_filters_test.dart
@@ -2,7 +2,6 @@ import 'package:equatable/equatable.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_meedu/flutter_meedu.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:flutter_meedu/router.dart' as router;
 
 void main() {
   setUp(() {