+++ packages/flutter_meedu/test/listeners/provider_listener_test.dart
@@ -1,9 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_meedu/flutter_meedu.dart';
-import 'package:flutter_meedu/meedu.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:flutter_meedu/router.dart' as router;
 
 import '../state/consumer_widget_test.dart';