+++ packages/flutter_meedu/test/listeners/provider_listener_test_with_filters_test.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_meedu/flutter_meedu.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:flutter_meedu/router.dart' as router;
 
 void main() {
   setUp(() {